requirements.txt
hume
asyncio
httpx
python-dotenv